Module Domain.Env

val empty : env

val from_vals : t Stdlib.Lazy.t Prelude.bwd -> int -> env

Create an environment from a list of values NOTE: This also takes in the length of the list to avoid performing an O(n) computation.

val lookup_idx : env -> int -> t Stdlib.Lazy.t option

Lookup a DeBrujin Index in an environment.

val lookup_tp_idx : env -> int -> tp option

Lookup a DeBrujin Index in an environment.

val locals : env -> t Stdlib.Lazy.t Prelude.bwd

Get all the local bindings in an environment.

val size : env -> int

Get the number of values in the environment.

val tp_size : env -> int

Get the number of types in the environment.

val extend : env -> t -> env

Extend an environment with a value.

val extend_tp : env -> tp -> env

Extend an environment with a type.