Module OrderedHashTbl.Make

Parameters

module H : Stdlib.Hashtbl.HashedType

Signature

type key = H.t

type 'a t

val create : int -> 'a t

Create a new, empty hashtable with a specified initial capacity.

val size : 'a t -> int

Return the number of entries in the hash table.

val push : key -> 'a -> 'a t -> unit

Add a new binding to the hash table, shadowing the previous bindings.

val pop : 'a t -> key * 'a

Remove the last binding from the table, potentially unshadowing.

val scope : key -> 'a -> 'a t -> ( unit -> 'r ) -> 'r

Bind a value to a key for the duration of a function.

val peek : 'a t -> key * 'a

Get the last binding in the table, but don't remove it.

val find : key -> 'a t -> 'a option

Get the most recently bound value for a given key.

val find_idx_of : key -> 'a t -> int option

Get the DeBruijin index of the most recently bound value for a given key.

val nth : int -> 'a t -> key * 'a

Get the nth last thing bound.

val values : 'a t -> 'a Bwd.bwd

Get all the values bound inside of the map.

val values_with : 'a t -> ( 'a -> 'b ) -> 'b Bwd.bwd

Get all the values bound inside of the map by using a projection function